renamed modules elias_fano and fast_rs_vec to ef and rs
