+++ src/trees/bp/mod.rs
@@ -3,7 +3,7 @@
 //! time, as well as subtree size, level-order, and ancestor queries in `O(log n)` time.
 //! The tree is succinct (ideally sublinear space overhead) and pointer-less.
 
-use crate::bit_vec::fast_rs_vec::SelectIntoIter;
+use crate::bit_vec::rs::SelectIntoIter;
 use crate::trees::mmt::MinMaxTree;
 use crate::trees::{IsAncestor, LevelTree, SubtreeSize, Tree};
 use crate::{BitVec, RsVec};
